Add FIleUploader component, that allows drag-and-drop and has a nicer upload behavior
Resize image before processing with tesseract

We still lose the preview file after upload, and it seems it is the old file that is not uploaded, not the resized file?

It still isn't pretty, but at least it resizes instantly, and that is what will be added to the generated pdf

Now has a modal that shows the uploaded PDF

The remove button currently doesn't work.

If I remove the explicit setting of value it does.
I assume it has something to do with controlled state.
The reason I set the value, is because otherwise the change isn't reflected when we resize the image.....

Fix the behavior with remove

Move the OCR into the FileUploader component